[BEAM-3734] Post code review fixes
