+++ sdks/java/io/common/src/test/java/org/apache/beam/sdk/io/common/TestRow.java
@@ -17,8 +17,6 @@
  */
 package org.apache.beam.sdk.io.common;
 
-import static org.apache.beam.sdk.io.common.IOTestHelper.getHashForRecordCount;
-
 import com.google.auto.value.AutoValue;
 import com.google.common.collect.ImmutableMap;
 import java.io.Serializable;
@@ -110,7 +108,10 @@ public abstract class TestRow implements Serializable, Comparable<TestRow> {
    */
   public static String getExpectedHashForRowCount(int rowCount)
       throws UnsupportedOperationException {
-
-    return getHashForRecordCount(rowCount, EXPECTED_HASHES);
+    String hash = EXPECTED_HASHES.get(rowCount);
+    if (hash == null) {
+      throw new UnsupportedOperationException("No hash for that row count");
+    }
+    return hash;
   }
 }
