+++ sdks/java/io/file-based-io-tests/pom.xml
@@ -349,6 +349,10 @@
             <artifactId>avro</artifactId>
             <scope>test</scope>
         </dependency>
-
+        <dependency>
+            <groupId>org.apache.beam</groupId>
+            <artifactId>beam-sdks-java-io-xml</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
