+++ sdks/java/io/xml/pom.xml
@@ -109,134 +109,9 @@
       <artifactId>hamcrest-library</artifactId>
       <scope>test</scope>
     </dependency>
-
-    <dependency>
-      <groupId>org.apache.beam</groupId>
-      <artifactId>beam-sdks-java-io-common</artifactId>
-      <scope>test</scope>
-      <classifier>tests</classifier>
-    </dependency>
-
-    <dependency>
-      <groupId>org.apache.beam</groupId>
-      <artifactId>beam-sdks-java-io-common</artifactId>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
 
   <profiles>
-    <!-- Include the Google Cloud Dataflow runner activated by -DintegrationTestRunner=dataflow -->
-    <profile>
-      <id>dataflow-runner</id>
-      <activation>
-        <property>
-          <name>integrationTestRunner</name>
-          <value>dataflow</value>
-        </property>
-      </activation>
-      <dependencies>
-        <dependency>
-          <groupId>org.apache.beam</groupId>
-          <artifactId>beam-runners-google-cloud-dataflow-java</artifactId>
-          <scope>runtime</scope>
-        </dependency>
-      </dependencies>
-    </profile>
-
-    <!--
-        This profile invokes PerfKitBenchmarker, which does benchmarking of
-        the IO ITs. The arguments passed to it allow it to invoke mvn again
-        with the desired benchmark.
-
-        To invoke this, run:
-
-        mvn verify -Dio-it-suite -pl sdks/java/io/xml
-            -DpkbLocation="path-to-pkb.py" \
-            -DintegrationTestPipelineOptions='["&ndash;&ndash;numberOfRecords=100000"]' \
-            -DioTest=fully-qualified IO IT name, eg. org.apache.beam.sdk.io.xml.XmlIOIT#writeThenReadViaSinkAndReadFiles
-
-        For DirectRunner, please use -DforceDirectRunner=true argument
-        For other runners please check doc in BEAM-3060 and https://beam.apache.org/documentation/io/testing/
-    -->
-    <profile>
-      <id>io-it-suite</id>
-      <activation>
-        <property><name>io-it-suite</name></property>
-      </activation>
-      <properties>
-        <!-- This is based on the location of the current pom relative to the root
-             See discussion in BEAM-2460 -->
-        <beamRootProjectDir>${project.parent.parent.parent.parent.basedir}</beamRootProjectDir>
-      </properties>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.codehaus.gmaven</groupId>
-            <artifactId>groovy-maven-plugin</artifactId>
-            <version>${groovy-maven-plugin.version}</version>
-            <executions>
-              <execution>
-                <id>find-supported-python-for-compile</id>
-                <phase>initialize</phase>
-                <goals>
-                  <goal>execute</goal>
-                </goals>
-                <configuration>
-                  <source>${beamRootProjectDir}/sdks/python/findSupportedPython.groovy</source>
-                </configuration>
-              </execution>
-            </executions>
-          </plugin>
-
-          <plugin>
-            <groupId>org.codehaus.mojo</groupId>
-            <artifactId>exec-maven-plugin</artifactId>
-            <version>${maven-exec-plugin.version}</version>
-            <executions>
-              <execution>
-                <phase>verify</phase>
-                <goals>
-                  <goal>exec</goal>
-                </goals>
-              </execution>
-            </executions>
-            <configuration>
-              <executable>${python.interpreter.bin}</executable>
-              <arguments>
-                <argument>${pkbLocation}</argument>
-                <argument>-benchmarks=beam_integration_benchmark</argument>
-                <argument>-beam_it_profile=io-it</argument>
-                <argument>-beam_it_timeout=${pkbTimeout}</argument>
-                <argument>-beam_location=${beamRootProjectDir}</argument>
-                <argument>-beam_prebuilt=true</argument>
-                <argument>-beam_sdk=java</argument>
-                <!-- runner overrides, controlled via forceDirectRunner -->
-                <argument>${pkbBeamRunnerProfile}</argument>
-                <argument>${pkbBeamRunnerOption}</argument>
-                <!-- specific to this IO -->
-                <argument>-beam_it_module=sdks/java/io/xml</argument>
-                <argument>-beam_it_class=${ioTest}</argument>
-                <!-- arguments typically defined by user -->
-                <argument>-beam_it_options=${integrationTestPipelineOptions}</argument>
-                <!--
-                optional array of key=value items. It will be passed to
-                target mvn command by pkb. eg. -DpkbExtraProperties='["filesystem=local"]'
-                -->
-                <argument>-beam_extra_mvn_properties=${pkbExtraProperties}</argument>
-              </arguments>
-            </configuration>
-          </plugin>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-surefire-plugin</artifactId>
-            <configuration>
-              <skipTests>true</skipTests>
-            </configuration>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
-
     <profile>
       <id>java-9</id>
       <activation>
